+++ tests/test_engine_required.py
@@ -41,6 +41,7 @@ def test_verify_orchestrator_preserves_required_policy_for_line_complexity_and_d
                 "line",
                 "cmake_lint",
                 "pyproject_lint",
+                "file_hygiene",
                 "lint",
                 "test",
                 "type",